pythonProject/requirements.txt
altgraph==0.17
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
future==0.18.2
idna==2.10
image==1.5.33
MouseInfo==0.1.3
numpy==1.20.0
pefile==2019.4.18
Pillow==8.1.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
pytz==2021.1
pywhatkit==3.9
pywin32-ctypes==0.2.0
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.3
wikipedia==1.4.0
